Extract common argument parsing
Makes it easier to share this code with the internal repo.
